fix: extracted migration logic to ApiKeyStorageMigration class
