+++ core/src/main/java/com/amplitude/core/utilities/FileStorage.kt
@@ -44,11 +44,11 @@ class FileStorage(
         }
     }
 
-    override fun write(key: Storage.Constants, value: String) {
+    override suspend fun write(key: Storage.Constants, value: String) {
         propertiesFile.putString(key.rawVal, value)
     }
 
-    override fun remove(key: Storage.Constants) {
+    override suspend fun remove(key: Storage.Constants) {
         propertiesFile.remove(key.rawVal)
     }
